requirements.txt
amqp==2.6.1
argcomplete==1.12.3
asgiref==3.5.2
billiard==3.6.4.0
celery==4.4.1
certifi==2022.6.15
charset-normalizer==2.1.0   
colorama==0.4.5
commitizen==2.29.4
decli==0.5.2
Django==3.2
django-celery-beat==2.0.0   
django-celery-results==1.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.4 
flower==0.9.5
humanize==4.3.0
idna==3.3
Jinja2==3.1.2
kombu==4.6.11
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.8.0    
prompt-toolkit==3.0.30
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomlkit==0.11.1
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.11
vine==1.3.0
wcwidth==0.2.5
